Changed everything to work with floats, we have a dispatcher setup that appears to be working properly, and message passing is working in one direction

+++ dummy_src/example.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int run_motor_fn(int motor_id, int pwm_frequency) {
+int run_motor_fn(float motor_id, float pwm_frequency) {
   printf("Hey you got it!");
   return 1;
 }
